// src/apis/generated/models/registerAdditionalInfoGender.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Project Name
 * 개발: [api.dev.my-domain.com](https://api.dev.my-domain.com)
<br/>
운영: [api.my-domain.com](https://api.my-domain.com)

 * OpenAPI spec version: 1.0.0
 */

/**
 * @nullable
 */
export type RegisterAdditionalInfoGender =
  | (typeof RegisterAdditionalInfoGender)[keyof typeof RegisterAdditionalInfoGender]
  | null

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RegisterAdditionalInfoGender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  NONE: 'NONE',
} as const